Move all the test processes in a single directory
