+++ py_kwiver/sprokit/tests/processes/examples.py
@@ -1,5 +1,5 @@
 #ckwg +28
-# Copyright 2012-2016 by Kitware, Inc.
+# Copyright 2012 by Kitware, Inc.
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -28,7 +28,7 @@
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 
-from sprokit.pipeline import process
+from kwiver.sprokit.pipeline import process
 
 
 class TestPythonProcess(process.PythonProcess):
@@ -36,16 +36,58 @@ class TestPythonProcess(process.PythonProcess):
         process.PythonProcess.__init__(self, conf)
 
 
-def __sprokit_register__():
-    from sprokit.pipeline import process_factory
+class PythonPrintNumberProcess(process.PythonProcess):
+    def __init__(self, conf):
+        process.PythonProcess.__init__(self, conf)
 
-    module_name = 'python:test.python.extras'
+        self.declare_configuration_key(
+            'output',
+            '',
+            'The path of the file to output to.')
+
+        flags = process.PortFlags()
+
+        flags.add(self.flag_required)
+
+        self.declare_input_port(
+            'input',
+            'integer',
+            flags,
+            'Where numbers are read from.')
+
+    def _configure(self):
+        path = self.config_value('output')
+
+        if not path:
+            raise RuntimeError('The path given was empty')
+
+        self.fout = open(path, 'w+')
+        self.fout.flush()
+
+        self._base_configure()
+
+    def _reset(self):
+        self.fout.close()
+
+        self._base_reset()
+
+    def _step(self):
+        num = self.grab_value_from_port('input')
+
+        self.fout.write('%d\n' % num)
+        self.fout.flush()
+
+        self._base_step()
+
+def __sprokit_register__():
+    from kwiver.sprokit.pipeline import process_factory
+
+    module_name = 'python:kwiver.sprokit.tests.processes'
 
     if process_factory.is_process_module_loaded(module_name):
         return
 
-    process_factory.add_process('extra_test_python_process',
-                                'An extra test Python process',
-                                TestPythonProcess)
+    process_factory.add_process('test_python_process', 'A test Python process', TestPythonProcess)
+    process_factory.add_process('pyprint_number', 'A Python process which prints numbers', PythonPrintNumberProcess)
 
     process_factory.mark_process_module_as_loaded(module_name)
