+++ py_kwiver/sprokit/util/CMakeLists.txt
@@ -29,3 +29,8 @@ kwiver_install_headers(
   SUBDIR   sprokit/python/util
   NOPATH
   )
+
+kwiver_create_python_init(sprokit/util)
+kwiver_add_python_module( ${CMAKE_CURRENT_SOURCE_DIR}/test.py
+                          sprokit/util
+                          test)
